requirements.txt
requests
python-dotenv
mysql-connector-python
python-dateutil
pandas
fastapi
uvicorn
pydantic
python-multipart
neuralprophet